requirements.txt
yt-dlp==2024.12.13
openai==1.93.2
pydantic==2.11.7
python-dotenv==1.0.1